Move snap A-cluster into SnapManager (pure refactor, no behavior change)
FrameManager decomposition cut #3. Pure move + call-site qualifier changes;
no behavior change.

- Move CascadeStack, GetFrameIdFromWindow, ShowSnapPreview, AnimateSnapConfirmation
  and 3 _snap* fields from FrameManager into SnapManager.
- Drop 8 `Framemanager.` reverse-call prefixes inside SnapManager — this
  eliminates the existing SnapManager <-> Framemanager circular dependency
  (those methods were always just callbacks for SnapManager anyway).
- Rewrite the 1 internal FrameManager call site (SizeChanged cascade hook)
  to SnapManager.CascadeStack(...).
- Add `using System.Windows.Controls;` to SnapManager (for Border).
- Delete zero-call dead code: IsFrameOrAncestorAutoRolled.

No visibility changes. FrameManager.cs: 9,963 -> 9,814 lines.
